# api-docs.yaml
{
  "openapi": "3.0.1",
  "info": {
    "title": "Demo Transaction Service",
    "description": "Demo transaction service api documentation",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Generated server url"
    }
  ],
  "paths": {
    "/transactions/transfer": {
      "post": {
        "tags": [
          "transaction-controller"
        ],
        "operationId": "makeTransfer",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TransferRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/TransferResponse"
                }
              }
            }
          }
        }
      }
    },
    "/transactions": {
      "get": {
        "tags": [
          "transaction-controller"
        ],
        "operationId": "findAllTransactions",
        "parameters": [
          {
            "name": "accountNumber",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "enum": [
                "PENDING",
                "SUCCESSFUL",
                "INSUFFICIENT_FUNDS",
                "FAILED",
                "INVALID_ACCOUNT_NUMBER",
                "INVALID_RECIPIENT"
              ]
            }
          },
          {
            "name": "startDate",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "format": "date"
            }
          },
          {
            "name": "endDate",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "format": "date"
            }
          },
          {
            "name": "category",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "enum": [
                "TRANSFERS",
                "BILL PAYMENTS",
                "AIRTIME"
              ]
            }
          },
          {
            "name": "channel",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "enum": [
                "MOBILE",
                "INTERNET",
                "USSD"
              ]
            }
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 1
            }
          },
          {
            "name": "size",
            "in": "query",
            "required": false,
            "schema": {
              "maximum": 100,
              "type": "integer",
              "format": "int32",
              "default": 10
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TransactionDTO"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/transactions/summary": {
      "get": {
        "tags": [
          "transaction-controller"
        ],
        "operationId": "generateTransactionSummary",
        "parameters": [
          {
            "name": "startDate",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "format": "date"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TransactionSummaryResponse"
                  }
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "TransferRequest": {
        "required": [
          "amount",
          "channel",
          "currency",
          "destinationAccountNumber",
          "paymentReference",
          "sourceAccountNumber",
          "type"
        ],
        "type": "object",
        "properties": {
          "paymentReference": {
            "maxLength": 80,
            "minLength": 0,
            "type": "string"
          },
          "amount": {
            "minimum": 1,
            "type": "number"
          },
          "channel": {
            "type": "string",
            "enum": [
              "MOBILE",
              "INTERNET",
              "USSD"
            ]
          },
          "description": {
            "maxLength": 1000,
            "minLength": 0,
            "type": "string"
          },
          "type": {
            "type": "string",
            "enum": [
              "TRANSFERS",
              "BILL PAYMENTS",
              "AIRTIME"
            ]
          },
          "currency": {
            "maxLength": 3,
            "minLength": 3,
            "type": "string"
          },
          "destinationAccountNumber": {
            "maxLength": 10,
            "minLength": 10,
            "type": "string"
          },
          "sourceAccountNumber": {
            "maxLength": 10,
            "minLength": 10,
            "type": "string"
          }
        }
      },
      "TransferResponse": {
        "type": "object",
        "properties": {
          "paymentReference": {
            "type": "string"
          }
        }
      },
      "TransactionDTO": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "accountNumber": {
            "type": "string"
          },
          "accountName": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "status": {
            "type": "string"
          },
          "category": {
            "type": "string"
          },
          "paymentReference": {
            "type": "string"
          },
          "channel": {
            "type": "string"
          },
          "currency": {
            "type": "string"
          },
          "amount": {
            "type": "number"
          },
          "date": {
            "type": "string",
            "format": "date-time"
          },
          "isCommissionWorthy": {
            "type": "boolean"
          },
          "statusMessage": {
            "type": "string"
          },
          "commission": {
            "type": "number"
          },
          "fee": {
            "type": "number"
          },
          "transactionEntries": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TransactionEntryDTO"
            }
          }
        }
      },
      "TransactionEntryDTO": {
        "type": "object",
        "properties": {
          "accountNumber": {
            "type": "string"
          },
          "accountName": {
            "type": "string"
          },
          "type": {
            "type": "string"
          },
          "amount": {
            "type": "number"
          }
        }
      },
      "TransactionSummaryResponse": {
        "type": "object",
        "properties": {
          "totalAmount": {
            "type": "number"
          },
          "channel": {
            "type": "string",
            "enum": [
              "MOBILE",
              "INTERNET",
              "USSD"
            ]
          },
          "transactionDate": {
            "type": "string",
            "format": "date"
          },
          "totalCommission": {
            "type": "number"
          },
          "type": {
            "type": "string",
            "enum": [
              "CREDIT",
              "DEBIT"
            ]
          }
        }
      }
    }
  }
}